Задание 3-4
деструктуризация
